requirements.txt
rdflib
sentence-transformers
faiss-cpu
numpy
torch
